backend/requirements.txt
fastapi
uvicorn
soundfile
numpy
praat-parselmouth
python-multipart
pydub